requirements.txt
streamlit
pandas
numpy
matplotlib
scikit-learn
tensorflow-cpu
joblib